requirements.txt
flask
requests
flask-pymongo
dnspython
google-generativeai
gradio
